+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog].
 <!-- next-header -->
 ## Unreleased - ReleaseDate
 
+## 0.13.13 - 2026-07-15
+
 ### Fixes
 
 - *(upgrade)* Point single-`--verbose` users to `--verbose --verbose` when more dependencies are hidden